consultant/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
pymysql
DBUtils
openai
httpx
openai-agents
pystun3
redis
orjson
PyJWT
requests
anyio
